:recycle: :: Number Type -> Specify Type

+++ core/network/src/main/java/com/school_of_company/network/mapper/post/response/PostListResponseMapper.kt
@@ -22,7 +22,7 @@ fun PostDto.toModel(): Post {
         mode = this.mode,
         title = this.title,
         content = this.content,
-        gwangsan = this.gwangsan.toInt(),
+        gwangsan = this.gwangsan,
         imageUrls = this.imageUrls.toModel()
     )
 }
